+++ paper.bib
@@ -1,627 +1,13 @@
-@PREAMBLE{
- "\providecommand{\noopsort}[1]{}" 
- # "\providecommand{\singleletter}[1]{#1}%" 
-}
-
-@article{schweizer1989microscopic,
-  title={Microscopic theory of the dynamics of polymeric liquids: General formulation of a mode--mode-coupling approach},
-  author={Schweizer, Kenneth S},
-  journal={J. Chem. Phys. },
-  volume={91},
-  number={9},
-  pages={5802--5821},
-  year={1989},
-  publisher={American Institute of Physics}
-}
-
-@article{mori1965transport,
-  title={Transport, collective motion, and Brownian motion},
-  author={Mori, Hazime},
-  journal={Prog. Theor. Phys.},
-  volume={33},
-  number={3},
-  pages={423--455},
-  year={1965},
-  publisher={Oxford University Press}
-}
-
-
-@article{weysser2010structural,
-  title={Structural relaxation of polydisperse hard spheres: Comparison of the mode-coupling theory to a Langevin dynamics simulation},
-  author={Weysser, Fabian and Puertas, Antonio Manuel and Fuchs, Matthias and Voigtmann, Th},
-  journal={Phys. Rev. E},
-  volume={82},
-  number={1},
-  pages={011504},
-  year={2010},
-  publisher={APS}
-}
-
-@book{dhont1996introduction,
-  title={An introduction to dynamics of colloids},
-  author={Dhont, Jan K G},
-  year={1996},
-  publisher={Elsevier}
-}
-
-@article{nagele1999cooperative,
-  title={Cooperative diffusion in colloidal mixtures},
-  author={N{\"a}gele, G and Bergenholtz, J and Dhont, J K G},
-  journal={J. Chem. Phys. },
-  volume={110},
-  number={14},
-  pages={7037--7052},
-  year={1999},
-  publisher={American Institute of Physics}
-}
-
-@article{coslovich2019localization,
-  title={A localization transition underlies the mode-coupling crossover of glasses},
-  author={Coslovich, Daniele and Ninarello, Andrea and Berthier, Ludovic},
-  journal={SciPost Phys.},
-  volume={7},
-  number={6},
-  pages={077},
-  year={2019}
-}
-
-@article{das2022crossover,
-  title={Crossover in dynamics in the Kob-Andersen binary mixture glass-forming liquid},
-  author={Das, Pallabi and Sastry, Srikanth},
-  journal={J. Non-Cryst. Solids X},
-  volume={14},
-  pages={100098},
-  year={2022},
-  publisher={Elsevier}
-}
-
-@article{andersen2005molecular,
-  title={Molecular dynamics studies of heterogeneous dynamics and dynamic crossover in supercooled atomic liquids},
-  author={Andersen, Hans C},
-  journal={Proc. Natl. Acad. Sci.},
-  volume={102},
-  number={19},
-  pages={6686--6691},
-  year={2005},
-  publisher={National Acad Sciences}
-}
-
-@article{bhattacharyya2008facilitation,
-  title={Facilitation, complexity growth, mode coupling, and activated dynamics in supercooled liquids},
-  author={Bhattacharyya, Sarika Maitra and Bagchi, Biman and Wolynes, Peter G},
-  journal={Proc. Natl. Acad. Sci.},
-  volume={105},
-  number={42},
-  pages={16077--16082},
-  year={2008},
-  publisher={National Acad Sciences}
-}
-
-@article{kob1995testing,
-  title={Testing mode-coupling theory for a supercooled binary Lennard-Jones mixture. II. Intermediate scattering function and dynamic susceptibility},
[…]
