### 2023-08-03 #5718 @mstimberg
:wave: @fverdugo & @epspebble, @ericneiva would any of you be willing to review this submission (_“ModeCouplingTheory.jl: A solver for mode-coupling-theory-like integro-differential equations”_) for [JOSS](joss.theoj.org/)? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
