### 2023-08-03 #5718 @epspebble
Hi Marcel,

Yes, it looks fun. I have reviewed the guidelines, quickly scanned the
GitHub repo of ModeCouplingTheory and the previous YellowBrick review
process.

Please let me know if we’d proceed with this and the time frame for
completing the checkboxes.

Cheers,
Simon

On Thu, Aug 3, 2023 at 1:27 PM Marcel Stimberg ***@***.***>
wrote:

> 👋 @fverdugo <https://github.com/fverdugo> & @epspebble
> <https://github.com/epspebble>, @ericneiva <https://github.com/ericneiva>
> would any of you be willing to review this submission (*“ModeCouplingTheory.jl:
> A solver for mode-coupling-theory-like integro-differential equations”*)
> for JOSS <http://joss.theoj.org/>? We carry out our checklist-driven
> reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
