### 2023-08-03 #5718 @mstimberg
Great, thanks for the quick reply, @epspebble ! I'll let you know when the review officially starts. In general, we aim to finish reviews at JOSS within a few weeks, but given that I will be on away vacation myself very soon, there's no rush :) If you could aim for doing the review until early/mid September, that would be perfect.

PS: when you say you locked at "the previous YellowBrick review process", I assume you mean you had a lock at openjournals/joss-reviews#1075? I did not check whether this review process has been particularly exemplary, but please be aware that there have been a number of  technical changes to the review process since then (e.g. `@editorialbot` replaced `@whedon`).
