### 2023-08-04 #5718 @epspebble
Concerning the process, it should be fine since I am new and it doesn’t
seem you have changed too much in the process. The bot idea is wonderful.
For timing, I would prefer to do the bulk of the work late August to early
September (which fits your schedule), but not much later than that as I
happen to teach 3 courses this Fall with some new administrative workload.

Feel free to send me a more “exemplary” thread of the review process. I
looked at YellowBricks because it was explicitly quoted somewhere in the
guidelines. So you might want to replace that soon if that doesn’t reflect
the current practice too well.

On Thu, Aug 3, 2023 at 3:32 PM Marcel Stimberg ***@***.***>
wrote:

> Great, thanks for the quick reply, @epspebble
> <https://github.com/epspebble> ! I'll let you know when the review
> officially starts. In general, we aim to finish reviews at JOSS within a
> few weeks, but given that I will be on away vacation myself very soon,
> there's no rush :) If you could aim for doing the review until early/mid
> September, that would be perfect.
>
> PS: when you say you locked at "the previous YellowBrick review process",
> I assume you mean you had a lock at #1075
> <https://github.com/openjournals/joss-reviews/issues/1075>? I did not
> check whether this review process has been particularly exemplary, but
> please be aware that there have been a number of technical changes to the
> review process since then (e.g. @editorialbot replaced @whedon).
>
