### 2023-08-04 #5718 @mstimberg
> For timing, I would prefer to do the bulk of the work late August to early September (which fits your schedule), but not much later than that as I happen to teach 3 courses this Fall with some new administrative workload.

Reviews for JOSS are very freely structured with respect to who does what when, so I am rather confident that this will work out.

> I looked at YellowBricks because it was explicitly quoted somewhere in the guidelines.
Thanks, I wasn't aware of that.
