### 2023-08-04 #5737 @editorialbot
[REVIEW]: ModeCouplingTheory.jl: A solver for mode-coupling-theory-like integro-differential equations
**Submitting author:** @IlianPihlajamaa (<a href="http://orcid.org/0000-0003-3779-4281">Ilian Pihlajamaa</a>)
**Repository:** https://github.com/IlianPihlajamaa/ModeCouplingTheory.jl
**Branch with paper.md** (empty if default branch):
**Version:** 0.8.1
