### 2023-09-01 #5737 @mstimberg
> HI @mstimberg - can you help me understand what's going on with this? I don't see any progress in the last month, and I only see 1 assigned reviewer, so I'm a little confused

Hi @danielskatz. I was away for all of August (see https://github.com/openjournals/joss-reviews/issues/5718#issuecomment-1664423068) and will carry things forward from next week on. I opened the review issue for the single reviewer that agreed so far so that they could already start, since they said that they'd have time late August/early September (https://github.com/openjournals/joss-reviews/issues/5718#issuecomment-1664842419) but might not be available later. As soon as I find someone, I'll add the second reviewer here.
