### 2023-09-07 #5737 @mstimberg
@epspebble how are looking things from your side, did you find time to look at the software/paper yet? Please be reminded that JOSS reviews can be iterative, i.e. feel free to open issues or comment here as soon as you come across concerns – there is no need to gather everything into a single review text as commonly done for traditional journals. In the same vein, please generate your checklist and feel free to tick boxes in your checklist whenever you are sure about them (the two in the beginning – conflict of interest and code of conduct – would be good first candidates 😉 ). Thank you for your time 🙏 !
