### 2023-10-03 #5737 @epspebble
Hi Marcel,

September has been very busy as I mentioned to you earlier. There’ll be a
big push coming as Canadian thanksgiving is around the corner!

Simon

On Mon, Oct 2, 2023 at 1:01 AM Marcel Stimberg ***@***.***>
wrote:

> 👋 @epspebble <https://github.com/epspebble>, @dawbarton
> <https://github.com/dawbarton> how are things going with the reviews?
> Please let me know if there is anything I can do from my side to help
> advance things 🙏
>
