### 2023-10-10 #5737 @epspebble
Hi Marcel,

Just double checking with you here concerning the bullet points under "Software paper".

IMO, the author has made some effort towards helping a "diverse, non-specialist" audience to understand how to use the software, but I think the software is made to solve a specialist class of problem, namely Mode Coupling Theory. Does the nature of this software allows us to forgive the lack of diversity? I do not see examples of integro-differential equations in the software paper that do not arise from Mode Coupling Theory (MCT). In fact, that's only one example, but that seem sophisticated enough for people working in MCT.

The next question is just a quick confirmation. I simply checked "State of the Field" because I am persuaded by the author that indeed there are no comparable software that is convenient to do this (except "in-house integrators"). The similar enough situation is discussed here: https://discourse.julialang.org/t/integro-differential-equation-and-2nd-order-differential-equation/78799/9 (which requires some extra setup)

So, can we just check this item whenever indeed there are no replacements? Do we need to do a more extensive search?

Simon
