### 2023-10-12 #5737 @dawbarton
Overall, this is a good submission and I'd be keen for this to be published in JOSS. I don't have a background in MCT and don't know the literature in this area, but the general approach/methods are well within my area of expertise. All the examples provided work in the latest version of the software and all tests pass. The documentation and paper are generally very well written.

My comments are all really minor. The only thing that is missing are the community guidelines - at least, I didn't spot any.

## Paper

- Minor point: the acronym MCT is used before it is introduced.
- Might be good to have a reference for Bengtzelius' trick (I'll leave that up to you to decide if you want to or not).
- Statement of need: point 5 is already mentioned in point 1. I'd just remove point 5 as the claims are sufficiently strong anyway. (Minor: point 7 "long-time limits of the equation" -> "long-time limits of the solution".)

## Documentation

- On the kernels page, one of the sections is numbered but nothing else is "1. `ExponentiallyDecayingKernel`"
