+++ paper.bib
@@ -10,6 +10,17 @@
   doi={10.1088/0022-3719/17/33/005}
 }
 
+@article{bengtzelius1986dynamics,
+  title={Dynamics of a Lennard-Jones system close to the glass transition},
+  author={Bengtzelius, U},
+  journal={Phys. Rev. A},
+  volume={34},
+  number={6},
+  pages={5059},
+  year={1986},
+  publisher={APS}
+}
+
 @article{berthier2013non,
   title={Non-equilibrium glass transitions in driven and active matter},
   author={Berthier, Ludovic and Kurchan, Jorge},
