### 2023-10-19 #5737 @IlianPihlajamaa
We would like to thank @dawbarton for going through the paper and the software documentation. Below we address their comments and suggestions.
-	There were some community guidelines at the bottom of the README document, but we agree that they were not easily discoverable. We have added them as a separate section now, both in the README as well as in the documentation.
-	We thank the referee for pointing out that MCT is not properly defined. We have corrected it in the manuscript.
-	As far as we know, the most detailed report in the literature is in section IV of https://doi.org/10.1103/PhysRevA.34.5059, where Bengtzelius writes that "By using a special technique of storing intermediate results from the inner integration the computational time increases linearly with the number of k values rather then quadratically." We have added this as a reference to the manuscript.
-	We agree and have removed point 5 (and corrected the sentence in current point 6).
-	We have fixed the formatting issue in the documentation.

Through our response and changes, we hope to have adequately addressed the comments.

@mstimberg What should I do to update the submitted manuscript, other than editing the paper.md file in our github repository?
