### 2023-10-19 #5737 @mstimberg
> What should I do to update the submitted manuscript, other than editing the paper.md file in our github repository?

You can make a comment asking `@editorialbot generate pdf`, and it will generate the new manuscript for you. Otherwise let me know and I can trigger it for you.
