### 2023-10-24 #5737 @epspebble
Not having a background in MCT, I am pleasantly impressed by the level of detail the authors put in the documentation to explain how to use the software - they are both succinct and precise. With their help, I have successfully walked through the examples, modifying them and re-run with slightly different sets of parameters with success. I also commend the abundance of references put at appropriate places, including function docstrings. I am happy to recommend acceptance of this work to JOSS.

I have not found any major errors except typos (e.g. succesful should be successful in the paper)
