+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ModeCouplingTheory"
 uuid = "6cad4188-63cb-4528-a10f-dbb409b456c6"
 authors = ["s158686 "]
-version = "0.8.0"
+version = "0.8.1"
 
 
 [deps]
