### 2023-10-24 #5737 @mstimberg
> @mstimberg I accidentally checked one of your tasks and now I cannot undo that.

@IlianPihlajamaa I don't see anything checked, so seems to be ok?
