### 2023-10-24 #5737 @IlianPihlajamaa
The software version is: 0.8.1.
The Zenodo doi is: [10.5281/zenodo.10036791](https://zenodo.org/doi/10.5281/zenodo.10036791).
