+++ v0.8.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"Kernels.html#Kernels","page":"Kernels","title":"Kernels","text":"","category":"section"},{"location":"Kernels.html","page":"Kernels","title":"Kernels","text":"A memory kernel kernel is an instance of a type of which MemoryKernel is a supertype. It can be used to compute out = evaluate_kernel(kernel, F, t) to find the value of the memory kernel. Additionally, when F is a mutable container like a Vector, one can call evaluate_kernel!(out, kernel, F, t) in which case it will mutate the elements of the temporary array out. Below we list the memory kernels that this package defines and give some examples of how to use them.","category":"page"},{"location":"Kernels.html#Schematic-Kernels","page":"Kernels","title":"Schematic Kernels","text":"","category":"section"},{"location":"Kernels.html","page":"Kernels","title":"Kernels","text":"This package includes a couple of schematic memory kernels.","category":"page"},{"location":"Kernels.html#ExponentiallyDecayingKernel","page":"Kernels","title":"ExponentiallyDecayingKernel","text":"","category":"section"},{"location":"Kernels.html","page":"Kernels","title":"Kernels","text":"The ExponentiallyDecayingKernel implements the kernel K(t) = λ exp(-tτ). It has fields λ <: Number and τ <: Number.","category":"page"},{"location":"Kernels.html#Example","page":"Kernels","title":"Example","text":"","category":"section"},{"location":"Kernels.html","page":"Kernels","title":"Kernels","text":"The integro-differential equation with this memory kernel actually has an analytic solution for tau=1, alpha=0 , beta=1, and gamma=1. It is given by","category":"page"},{"location":"Kernels.html","page":"Kernels","title":"Kernels","text":"F(t) = frace^-fract2left( lambda + sqrtlambda(lambda+4) + 2right)2 sqrtlambda  (lambda +4)left(sqrtlambda(lambda+4) left(e^sqrtlambda(lambda+4) t+1right)+lambda  left(e^sqrtlambda(lambda+4) t-1right)right)","category":"page"},{"location":"Kernels.html","page":"Kernels","title":"Kernels","text":"F0 = 1.0; ∂F0 = 0.0; α = 0.0; β = 1.0; γ = 1.0; δ = 0.0; λ = 1.0; τ = 1.0;\n\nkernel = ExponentiallyDecayingKernel(λ, τ)\nproblem = MemoryEquation(α, β, γ, δ, F0, ∂F0, kernel)\nsolver = TimeDoublingSolver(Δt=10^-3, t_max=10.0^2, verbose=false, N = 128, tolerance=10^-10, max_iterations=10^6)\nsol =  solve(problem, solver)\n\nt_analytic = 10 .^ range(-3, 2, length=50)\nF_analytic = @. (exp(-0.5*(3+sqrt(5))* t_analytic)*(exp(sqrt(5)*t_analytic) * (1+sqrt(5))-1+sqrt(5)))/(2sqrt(5))\n\nusing Plots\np = plot(log10.(get_t(sol)), get_F(sol), label=\"Numeric solution\", lw=3)\nscatter!(log10.(t_analytic), F_analytic, label=\"Exact solution\", ylabel=\"F\", xlabel=\"log10(t)\")","category":"page"},{"location":"Kernels.html","page":"Kernels","title":"Kernels","text":"(Image: image)","category":"page"},{"location":"Kernels.html#SchematicF1Kernel","page":"Kernels","title":"SchematicF1Kernel","text":"","category":"section"},{"location":"Kernels.html","page":"Kernels
[…]
