### 2023-11-02 280a4ce
Fix capitalization in bibliography and add a DOI
