+++ paper.bib
@@ -11,14 +11,15 @@
 }
 
 @article{bengtzelius1986dynamics,
-  title={Dynamics of a Lennard-Jones system close to the glass transition},
+  title={Dynamics of a {Lennard-Jones} system close to the glass transition},
   author={Bengtzelius, U},
   journal={Phys. Rev. A},
   volume={34},
   number={6},
   pages={5059},
   year={1986},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevA.34.5059}
 }
 
 @article{berthier2013non,
@@ -159,7 +160,7 @@
 }
 
 @article{winter2023deep,
-  title={A deep learning approach to the measurement of long-lived memory kernels from generalized Langevin dynamics},
+  title={A deep learning approach to the measurement of long-lived memory kernels from generalized {Langevin} dynamics},
   author={Kerr Winter, Max and Pihlajamaa, Ilian and Debets, Vincent E and Janssen, L M C},
   journal={J. Chem. Phys.},
   volume={158},
@@ -216,7 +217,7 @@
 
 
 @article{lawson1979basic,
-  title={Basic linear algebra subprograms for Fortran usage},
+  title={Basic linear algebra subprograms for {Fortran} usage},
   author={Lawson, Chuck L and Hanson, Richard J. and Kincaid, David R and Krogh, Fred T.},
   journal={ACM Trans. Math. Softw.},
   volume={5},
@@ -240,7 +241,7 @@
 }
 
 @article{liluashvili2017mode,
-  title={Mode-coupling theory for active Brownian particles},
+  title={Mode-coupling theory for active {Brownian} particles},
   author={Liluashvili, Alexander and {\'O}nody, Jonathan and Voigtmann, Thomas},
   journal={Phys. Rev. E},
   volume={96},
@@ -272,7 +273,7 @@
 }
 
 @article{reichert2021mode,
-  title={Mode-coupling theory for tagged-particle motion of active Brownian particles},
+  title={Mode-coupling theory for tagged-particle motion of active {Brownian} particles},
   author={Reichert, Julian and Mandal, Suvendu and Voigtmann, Thomas},
   journal={Phys. Rev. E},
   volume={104},
@@ -344,7 +345,7 @@
 }
 
 @article{wertheim1963exact,
-  title={Exact solution of the Percus-Yevick integral equation for hard spheres},
+  title={Exact solution of the {Percus-Yevick} integral equation for hard spheres},
   author={Wertheim, MS},
   journal={Phys. Rev. Lett.},
   volume={10},
@@ -356,7 +357,7 @@
 }
 
 @article{weysser2010structural,
-  title={Structural relaxation of polydisperse hard spheres: Comparison of the mode-coupling theory to a Langevin dynamics simulation},
+  title={Structural relaxation of polydisperse hard spheres: Comparison of the mode-coupling theory to a {Langevin} dynamics simulation},
   author={Weysser, Fabian and Puertas, Antonio Manuel and Fuchs, Matthias and Voigtmann, Th},
   journal={Phys. Rev. E},
   volume={82},
