### 2023-11-02 #5737 @mstimberg
Hi @IlianPihlajamaa thanks for checking off the items in the list, and apologies for taking so long to do the final review. I've opened a PR that fixes a few minor things (mostly capitalization in the bibliography) in IlianPihlajamaa/ModeCouplingTheory.jl#42. Apart from that, all is good from my side. Note that you do not have to release a new version and create a new archive, given that the changes only apply to the paper. As soon as this is done, I'll recommend the acceptance and trigger the final approval by the editor-in-chief.
