### 2023-11-03 #5737 @danielskatz
👋 @IlianPihlajamaa - I'm the track editor who will finish the processing of this submission. I've proofread the paper, and have a couple of small changes to request: https://github.com/IlianPihlajamaa/ModeCouplingTheory.jl/pull/43  Please merge this, or let me know what you disagree with, then we can continue the acceptance and publication process.
